EIP779 (DAO HF meta EIP): add the `extraData` requirement (#2911)
EIP779 (DAO HF meta EIP) "summarizes" what the DAO fork should do. However, it does not state that the extraData field of the blocks 1920000-1920009 should must be "dao-hard-fork". If someone wishes to implement their own Ethereum client, this is very important, because this gets explicitly tested in ethereum/tests and thus should be put in the "DAO fork summary" EIP.